requirements.txt
# runtime
pycryptodome

# dev
flake8
coverage
moto[server]
pytest
pytest-xprocess
pylint
pytest
